Don't allow adding to lists from attributes.
[Collection syntax](https://learn.microsoft.com/en-us/dotnet/desktop/wpf/advanced/xaml-syntax-in-detail?view=netframeworkdesktop-4.8#collection-syntax)
in XAML should only be supported when setting property values using property element syntax and not attribute syntax.

Adds a property to `XamlAstXamlPropertyValueNode` to indicate whether the property value came from an attribute,
and a property to `PropertySetterBinderParameters` to indicate whether a setter can be used from an attribute.

The `AdderSetter`s set `PropertySetterBinderParameters.AllowAttributeSyntax` to false, which disallows their usage
from attribute `XamlAstXamlPropertyValueNode`s.